requirements.txt

streamlit==1.31.0
langchain==0.1.0
langchain-groq==0.0.1
langchain-community==0.0.20
faiss-cpu==1.7.4
sentence-transformers==2.3.1
pypdf==4.0.1
python-dotenv==1.0.0